# gh-63207: Use GetSystemTimePreciseAsFileTime() in time.time()