--- a/Misc/NEWS.d/next/Library/2024-03-14-17-21-25.gh-issue-63207.LV16SL.rst
+++ b/Misc/NEWS.d/next/Library/2024-03-14-17-21-25.gh-issue-63207.LV16SL.rst
@@ -1,0 +1,4 @@
+On Windows, :func:`time.time()` now uses the
+``GetSystemTimePreciseAsFileTime()`` clock to have a resolution better than 1
+us, instead of the ``GetSystemTimeAsFileTime()`` clock which has a resolution
+of 15.6 ms. Patch by Victor Stinner.