@vstinner 2024-03-14
cc @pitrou @eryksun @zooba