@vstinner 2024-03-15
Clock resolution measured by the script: https://github.com/python/cpython/pull/116781#issuecomment-1997815440

Before:

```
time():
- efffective resolution: 15 ms
- implementation: GetSystemTimeAsFileTime()
- resolution: 16 ms

perf_counter():
- efffective resolution: 100 ns
- implementation: QueryPerformanceCounter()
- resolution: 100 ns

monotonic():
- efffective resolution: 100 ns
- implementation: QueryPerformanceCounter()
- resolution: 100 ns
```

After:

```
time():
- efffective resolution: 238 ns
- implementation: GetSystemTimePreciseAsFileTime()
- resolution: 100 ns

perf_counter():
- efffective resolution: 100 ns
- implementation: QueryPerformanceCounter()
- resolution: 100 ns

monotonic():
- efffective resolution: 100 ns
- implementation: QueryPerformanceCounter()
- resolution: 100 ns
```

`time.time()` resolution is way better!  238 ns instead of 15.6 ms, it's around 65 500x more accurate!