@pitrou 2024-03-15
> `time.time()` resolution is way better!  238 ns instead of 15.6 ms, it's around 65 500x more accurate!

Impressive! Finally Windows will be as good as other platforms here.